+++ .github/workflows/ci.yml
@@ -1,35 +1,20 @@
 name: CI
 
 on:
-    push:
+  push:
 
 jobs:
-    ci:
-        runs-on: ubuntu-latest
-        name: Code analysis and tests
+  ci:
+    runs-on: ubuntu-latest
+    name: Code analysis and tests
 
-        steps:
-            - uses: actions/checkout@v4
+    steps:
+      - uses: actions/checkout@v4
 
-            - name: Set up Python
-              uses: actions/setup-python@v5
-              with:
-                  python-version: "3.11"
+      - name: Setup .NET
+        uses: actions/setup-dotnet@v4
+        with:
+          dotnet-version: 8.0.x
 
-            - name: Install dependencies
-              run: |
-                  python -m pip install --upgrade pip
-                  pip install pylint black
-                  pip install -r requirements.txt
-
-            - name: Lint with pylint
-              run: |
-                  pylint $(git ls-files '*.py')
-
-            - name: Check formatting with black
-              run: |
-                  black $(git ls-files '*.py') --check --diff
-
-            - name: Tests
-              run: |
-                  python -m unittest discover -s test -p "test_*.py"
+      - name: Build solution
+        run: dotnet build BDMS.sln -c Release /warnaserror
