+++ Geodatenbezug.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34723.18
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TopicProcessing", "TopicProcessing.csproj", "{CAFFEF4B-AB45-47C9-88AA-30D98D872179}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Geodatenbezug", "Geodatenbezug\Geodatenbezug.csproj", "{FDF6653B-8A85-4E5A-97E8-A0097074CA5B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CAFFEF4B-AB45-47C9-88AA-30D98D872179}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CAFFEF4B-AB45-47C9-88AA-30D98D872179}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CAFFEF4B-AB45-47C9-88AA-30D98D872179}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CAFFEF4B-AB45-47C9-88AA-30D98D872179}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FDF6653B-8A85-4E5A-97E8-A0097074CA5B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FDF6653B-8A85-4E5A-97E8-A0097074CA5B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FDF6653B-8A85-4E5A-97E8-A0097074CA5B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FDF6653B-8A85-4E5A-97E8-A0097074CA5B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
