--- TopicProcessing/Properties/launchSettings.json
@@ -1,9 +1,0 @@
-{
-  "profiles": {
-    "TopicProcessing": {
-      "commandName": "Project",
-      "commandLineArgs": "--port 7295",
-      "launchBrowser": false
-    }
-  }
-}
