--- TopicProcessing/Properties/serviceDependencies.json
@@ -1,11 +1,0 @@
-{
-  "dependencies": {
-    "appInsights1": {
-      "type": "appInsights"
-    },
-    "storage1": {
-      "type": "storage",
-      "connectionId": "AzureWebJobsStorage"
-    }
-  }
-}
